Move small buffers tests from unit tests to integration tests
Those tests does not depend on the internals, so can be tested outside.
This will reduce the size of pack on crates.io
